pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "coastfi-navigator"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "dash>=2.17",
    "dash-mantine-components>=0.14",
    "plotly>=5.20",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    "numpy>=1.26",
    "scipy>=1.12",
    "bcrypt>=4.1",
    "python-dotenv>=1.0",
    "flask-login>=0.6",
]

[tool.uv]
dev-dependencies = [
    "hypothesis>=6.100",
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "black>=24.0",
    "ruff>=0.4",
]

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
